+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-Nothing so far
+- Add ReceiveMessageEvent.Data field to allow using the underlying message type of the adapters
+- Add ReceiveMessageEvent.AuthorID field to identify the author of the message
+- Add Message.Data field which contains a copy of the ReceiveMessageEvent.Data value
+- Add Message.AuthorID field which contains a copy of the ReceiveMessageEvent.AuthorID value 
 
 ## [v0.6.0] - 2019-03-30
 ### Added